gendered-pronoun-resolution/requirements.txt
absl-py==0.7.0
alabaster==0.7.12
allennlp==0.8.2
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.3.0
attrs==18.2.0
autopep8==1.4.3
aws-xray-sdk==0.95
awscli==1.16.123
Babel==2.6.0
backcall==0.1.0
bleach==3.1.0
boto==2.49.0
boto3==1.9.97
botocore==1.12.113
bz2file==0.98
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
chicksexer==0.2.2
Click==7.0
colorama==0.3.9
conllu==0.11
cookies==2.2.1
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.3.2
defusedxml==0.5.0
Deprecated==1.2.4
dill==0.2.9
docker==3.7.0
docker-pycreds==0.4.0
docutils==0.14
ecdsa==0.13
editdistance==0.5.3
entrypoints==0.3
flair==0.4.0
flake8==3.7.5
flaky==3.5.3
Flask==1.0.2
Flask-Cors==3.0.7
ftfy==5.5.1
future==0.17.1
gast==0.2.2
gensim==3.4.0
gevent==1.3.6
greenlet==0.4.15
grpcio==1.18.0
h5py==2.9.0
hyperopt==0.1.1
idna==2.8
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.3
jsondiff==1.1.1
jsonnet==0.10.0
jsonpickle==1.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
more-itertools==6.0.0
moto==1.3.4
mpld3==0.3
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
mypy==0.670
mypy-extensions==0.4.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.4
notebook==5.7.6
numpy==1.16.1
numpydoc==0.8.0
overrides==1.9
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
parsimonious==0.8.0
parso==0.3.4
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
plac==0.9.6
pluggy==0.8.1
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
ptyprocess==0.6.0
py==1.7.0
pyaml==18.11.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.7.3
pyflakes==2.1.0
Pygments==2.3.1
pymongo==3.7.2
pyparsing==2.3.1
pyrsistent==0.14.11
pytest==4.2.1
python-dateutil==2.8.0
python-jose==2.0.2
python-Levenshtein==0.12.0
pytorch-pretrained-bert==0.6.1
pytz==2017.3
PyYAML==3.13
pyzmq==18.0.1
qtconsole==4.4.3
regex==2019.2.18
requests==2.21.0
responses==0.10.5
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.20.2
scipy==1.2.1
seaborn==0.9.0
segtok==1.5.7
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.8.0
snowballstemmer==1.2.1
spacy==2.0.18
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
sqlitedict==1.6.0
sqlparse==0.2.4
syntok==1.1.0
tensorboard==1.12.2
tensorboardX==1.2
tensorflow==1.13.0rc2
tensorflow-estimator==1.13.0rc0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thinc==6.12.1
toolz==0.9.0
torch==1.0.0
tornado==6.0.2
tqdm==4.26.0
traitlets==4.3.2
typed-ast==1.3.1
typing==3.6.4
ujson==1.35
Unidecode==1.0.23
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.55.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xmltodict==0.12.0
